% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{arrange.disk_df}
\alias{arrange.disk_df}
\title{Arrange rows}
\usage{
\method{arrange}{disk_df}(.data, ...)
}
\arguments{
\item{.data}{A disk_df.}

\item{...}{Columns to order by. Use \code{\link[dplyr:desc]{desc()}} for descending order.}
}
\value{
A disk_df with the arrange operation queued.
}
\description{
Arrange rows
}
\examples{
df <- disk_df(mtcars)
collect(arrange(df, desc(mpg)))
}
