% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disk_df.R
\name{as.disk_df}
\alias{as.disk_df}
\title{Coerce to a disk_df}
\usage{
as.disk_df(x, ...)
}
\arguments{
\item{x}{An object to coerce.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A \code{disk_df}.
}
\description{
Coerce to a disk_df
}
