% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{glimpse.disk_df}
\alias{glimpse.disk_df}
\title{Glimpse at the structure}
\usage{
\method{glimpse}{disk_df}(x, ...)
}
\arguments{
\item{x}{A disk_df.}

\item{...}{Not used.}
}
\description{
Glimpse at the structure
}
