% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{group_by.disk_df}
\alias{group_by.disk_df}
\title{Group data for subsequent summarise}
\usage{
\method{group_by}{disk_df}(.data, ...)
}
\arguments{
\item{.data}{A disk_df.}

\item{...}{Column names to group by.}
}
\value{
A disk_df with grouping recorded.
}
\description{
Group data for subsequent summarise
}
\examples{
df <- disk_df(mtcars)
collect(df |> group_by(cyl) |> summarise(avg_mpg = mean(mpg)))
}
