% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disk_df.R
\name{is.disk_df}
\alias{is.disk_df}
\title{Test if an object is a disk_df}
\usage{
is.disk_df(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
`TRUE` if the object inherits from `disk_df`.
}
\description{
Test if an object is a disk_df
}
