% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_safe.R
\name{memory_safe_check}
\alias{memory_safe_check}
\title{Check and warn if a collect operation would load too much data}
\usage{
memory_safe_check(x, context = "operation", ...)
}
\arguments{
\item{x}{A disk_df.}

\item{context}{A string describing the calling function.}

\item{...}{Not used.}
}
\description{
Check and warn if a collect operation would load too much data
}
