% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_safe.R
\name{memory_safe_get_limit}
\alias{memory_safe_get_limit}
\title{Get current memory-safe limits}
\usage{
memory_safe_get_limit()
}
\value{
A list with the current \code{limit_rows}, \code{limit_bytes}, and \code{action}.
}
\description{
Get current memory-safe limits
}
