% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disk_df.R
\name{names.disk_df}
\alias{names.disk_df}
\title{Get or set column names}
\usage{
\method{names}{disk_df}(x)
}
\arguments{
\item{x}{A disk_df object.}

\item{value}{Character vector of names.}
}
\description{
Get or set column names
}
