% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{pull.disk_df}
\alias{pull.disk_df}
\title{Pull a single column}
\usage{
\method{pull}{disk_df}(.data, var = -1, ...)
}
\arguments{
\item{.data}{A disk_df.}

\item{var}{The column name or index.}

\item{...}{Not used.}
}
\value{
A vector.
}
\description{
Pull a single column
}
