% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{select.disk_df}
\alias{select.disk_df}
\title{Choose columns}
\usage{
\method{select}{disk_df}(.data, ...)
}
\arguments{
\item{.data}{A disk_df.}

\item{...}{One or more column names (unquoted).}
}
\value{
A disk_df with the select operation queued.
}
\description{
Choose columns
}
\examples{
df <- disk_df(mtcars)
collect(select(df, mpg, wt))
}
