% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{ungroup.disk_df}
\alias{ungroup.disk_df}
\title{Remove grouping}
\usage{
\method{ungroup}{disk_df}(.data, ...)
}
\arguments{
\item{.data}{A disk_df.}
}
\value{
A disk_df with grouping removed.
}
\description{
Remove grouping
}
